html_root/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
manim
numpy
python-multipart
supabase
bcrypt
Pillow